refactor: reduce survey complexity gates
